Operation Snake Oil: Rename PircBot package and remove unused code. Refs #210.
Some heuristics of "virus" scanner apps think that the existance of PircBot is an
indicator for malware. This is super annoying and a bad false positive.

This patch moves PircBot to the org.yaaic.protocol packagae and removes a bunch
of unused code. As part of this I also removed the untested, not maintainted
DCC send feature.

AVG is the last scanner that still thinks Yaaic is malware. More testing will
be needed to "fix" that.